+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Resolves an issue where using the transformer with `ts.transform` would throw an
 
 The transformer will no longer emit members with the `declare` modifier.  [stefan-lacatus](https://github.com/stefan-lacatus))
 
+Resolves an issue where the `__values` helper was in some cases not inlined, preventing the use of transpiled es6 features.
+
 # 1.3.1
 
 Resolves an issue where inline SQL statements would compile into code with syntax errors.